Implements draft creation for entities
Enables draft creation for entities by invoking a "CRUD_Create" method if it exists.
This allows for creating associated data or triggering events when an entity is created as a draft.

Adds a `Plus` method to the entity's node to handle child entity creation.
This method sets the child entity as a draft, adds it to the parent, calls create and sets it back to not draft.

Updates folder creation in Package.cs and adds basic CRUD logic for folders to use.
Updates the assembly definition file.