Adds metadata support to checkout sessions
Introduces metadata to checkout sessions.

The `Purchase` method in `CheckoutAPI` now accepts a `metadata` dictionary,
which is then passed to the current checkout session.
This metadata is included when creating charges,
allowing for more detailed transaction information.

Additionally, the `RunCheckoutSession` method is split into `CreateCheckoutSession`
and `RunCheckoutSession`, improving the flow and allowing injecting the metadata.